moved the logic of the database topic to src/db
